+++ eslint.config.utils.js
@@ -1,6 +1,8 @@
 import pluginJs from '@eslint/js';
 import eslintConfigPrettier from 'eslint-config-prettier';
 import eslintPluginJest from 'eslint-plugin-jest';
+import eslintPluginJestDom from 'eslint-plugin-jest-dom';
+import eslintPluginPrettier from 'eslint-plugin-prettier/recommended';
 import eslintPluginUnicorn from 'eslint-plugin-unicorn';
 import globals from 'globals';
 import tseslint from 'typescript-eslint';
@@ -29,13 +31,16 @@ function setUnicornAllowList(list = []) {
 export function setEslintPluginUnicorn({ rules, allowList } = {}) {
   return {
     ...eslintPluginUnicorn.configs['flat/recommended'],
-
     rules: {
+      ...eslintPluginUnicorn.configs['flat/recommended'].rules,
       'unicorn/switch-case-braces': ['error', 'avoid'],
+      'unicorn/prefer-logical-operator-over-ternary': 'off',
       'unicorn/prefer-string-replace-all': 'off',
       'unicorn/no-nested-ternary': 'off',
       'unicorn/no-array-reduce': 'off',
+      'unicorn/prefer-at': 'off',
       'unicorn/no-null': 'off',
+      'unicorn/filename-case': ['error', { cases: { kebabCase: true, pascalCase: true } }],
       'unicorn/prevent-abbreviations': ['error', setUnicornAllowList(allowList)],
       ...rules,
     },
@@ -47,9 +52,18 @@ export function setEslintPluginUnicorn({ rules, allowList } = {}) {
 export function setEslintPluginJest({ rules } = {}) {
   return {
     ...eslintPluginJest.configs['flat/recommended'],
-    files: ['tests/**'],
     rules: {
       ...eslintPluginJest.configs['flat/recommended'].rules,
+      ...rules,
+    },
+  };
+}
+
+export function setEslintPluginJestDom({ rules } = {}) {
+  return {
+    ...eslintPluginJestDom.configs['flat/recommended'],
+    rules: {
+      ...eslintPluginJestDom.configs['flat/recommended'].rules,
       ...rules,
     },
   };
@@ -58,12 +72,18 @@ export function setEslintPluginJest({ rules } = {}) {
 //
 
 export function setEslintPluginPrettier({ rules } = {}) {
-  return { ...eslintConfigPrettier, rules: { ...rules } };
+  return {
+    ...eslintPluginPrettier,
+    rules: {
+      ...eslintConfigPrettier.rules,
+      ...rules,
+    },
+  };
 }
 
 //
 
-export function setEslintPluginTypescripEslint({ rules } = {}) {
+export function setEslintPluginTypescriptEslint({ rules } = {}) {
   return [
     { ...pluginJs.configs.recommended },
     ...tseslint.configs.recommended,